refactor(ReservationServiceTest): 예약 날짜 예외처리를 domain -> service 로 옮김에 따라 테스트 메서드도 이동
